Register metrics in server start and move sender metrics to sender file
